02-adk-deployments/adk-deployment-01-agentengine-sdk/02_agent_with_mcp_toolset/requirements.txt
fastmcp==2.12.5
python-dotenv
google-cloud-aiplatform==1.126.1
google-adk==1.18.0